Moving utils/exceptions to shared dir & creating new schema file
Since we've created a new shared folder, lets move utils and
exceptions to this dir.

Also split all the schema code into a separate file from utils.

Multiple files affected by these changes due to changes in the import
clauses.
