+++ suzieq/cli/sqcmds/command.py
@@ -12,7 +12,7 @@ from prompt_toolkit import prompt
 from natsort import natsort_keygen
 from colorama import Fore, Style
 
-from suzieq.exceptions import UserQueryError
+from suzieq.shared.exceptions import UserQueryError
 
 
 @dataclass
