+++ suzieq/poller/services/bgp.py
@@ -5,8 +5,8 @@ from datetime import datetime
 from copy import deepcopy
 
 from suzieq.poller.services.service import Service
-from suzieq.utils import get_timestamp_from_cisco_time
-from suzieq.utils import get_timestamp_from_junos_time
+from suzieq.shared.utils import get_timestamp_from_cisco_time
+from suzieq.shared.utils import get_timestamp_from_junos_time
 
 
 class BgpService(Service):
