+++ suzieq/shared/exceptions.py
